+++ src/kernels/execution/cellExecutionMessageHandler.ts
@@ -32,9 +32,8 @@ import { disposeAllDisposables } from '../../platform/common/helpers';
 import { traceError, traceWarning } from '../../platform/logging';
 import { RefBool } from '../../platform/common/refBool';
 import { IDisposable, IExtensionContext } from '../../platform/common/types';
-import { concatMultilineString, isJupyterNotebook } from '../../platform/common/utils';
+import { concatMultilineString, formatStreamText, isJupyterNotebook } from '../../platform/common/utils';
 import { traceCellMessage, cellOutputToVSCCellOutput, translateCellDisplayOutput } from './helpers';
-import { formatStreamText } from '../../webviews/webview-side/common';
 import { swallowExceptions } from '../../platform/common/utils/decorators';
 import { noop } from '../../platform/common/utils/misc';
 import { ITracebackFormatter } from '../../kernels/types';
