+++ src/webviews/webview-side/common/index.ts
@@ -37,37 +37,6 @@ export function stripComments(str: string): string {
     return result;
 }
 
-// Took this from jupyter/notebook
-// https://github.com/jupyter/notebook/blob/b8b66332e2023e83d2ee04f83d8814f567e01a4e/notebook/static/base/js/utils.js
-// Remove characters that are overridden by backspace characters
-function fixBackspace(txt: string) {
-    let tmp = txt;
-    do {
-        txt = tmp;
-        // Cancel out anything-but-newline followed by backspace
-        tmp = txt.replace(/[^\n]\x08/gm, '');
-    } while (tmp.length < txt.length);
-    return txt;
-}
-
-// Remove chunks that should be overridden by the effect of
-// carriage return characters
-// From https://github.com/jupyter/notebook/blob/master/notebook/static/base/js/utils.js
-function fixCarriageReturn(txt: string) {
-    txt = txt.replace(/\r+\n/gm, '\n'); // \r followed by \n --> newline
-    while (txt.search(/\r[^$]/g) > -1) {
-        var base = txt.match(/^(.*)\r+/m)![1];
-        var insert = txt.match(/\r+(.*)$/m)![1];
-        insert = insert + base.slice(insert.length, base.length);
-        txt = txt.replace(/\r+.*$/m, '\r').replace(/^.*\r/m, insert);
-    }
-    return txt;
-}
-export function formatStreamText(str: string): string {
-    // Do the same thing jupyter is doing
-    return fixCarriageReturn(fixBackspace(str));
-}
-
 export function appendLineFeed(arr: string[], eol: string = '\n', modifier?: (s: string) => string) {
     return arr.map((s: string, i: number) => {
         const out = modifier ? modifier(s) : s;
